Refactor the human to cron stuff into a separate file and use json_object OpenAI response format
